backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI-powered resume screening and JD matching backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "apscheduler>=3.10.4",
    "email-validator>=2.3.0",
    "fastapi>=0.135.2",
    "google-generativeai>=0.8.6",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.48",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
